# src/main/resources/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Motor de Crédito API
  description: API para avaliação de crédito, consulta de saldo/score e campanhas de renegociação
  version: 1.0.0
servers:
  - url: /api/v1
...
paths:
  /avaliar/{id}:
    post:
      summary: Avaliar crédito de um cliente
      operationId: avaliarCredito
      tags:
        - motor-liberacao-credito
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: valorSolicitado
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Resultado da avaliação
          content:
            application/json:
              schema:
                type: boolean
                example: true
        '400':
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Cliente não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-04-03T11:30:00Z"
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Cliente não encontrado"
        path:
          type: string
          example: "/api/credito/cliente/123"